move the workload ref get function to the common lib
